Fix some clang-tidy errors

+++ util.h
@@ -12,7 +12,7 @@ static inline int nanompi_get_dtype_size(MPI_Datatype dtype)
     return dtype.size;
 }
 
-static inline size_t nanompi_get_msg_size(MPI_Datatype dtype, int count)
+static inline int nanompi_get_msg_size(MPI_Datatype dtype, int count)
 {
     return count * nanompi_get_dtype_size(dtype);
 }
